refactor(bootstrap): split agents-bootstrap-github into gh-preflight + summary (refs #3349)
Extract the gh-version arithmetic and gh/runtime-dep preflight
(parseGhVersion, compareSemver, preflightGh, preflightRuntimeDeps,
MIN_GH_VERSION) into lib/bootstrap/gh-preflight.js, and the result
formatting helpers (format*Summary, printSummary) into
lib/bootstrap/summary.js. agents-bootstrap-github.js is reduced to
bootstrap orchestration that imports the two new modules and re-exports
the preflight surface for existing test consumers.
